Moved tree generation functions to tree.py
